requirements.txt
flask
qrcode[pil]
supabase
gunicorn
python-dotenv
flask-limiter
pytest